+++ bin/docspec
@@ -8,16 +8,16 @@ include Colsole
 
 document = ARGV[0] || 'README.md'
 abort "File not found #{document}" unless File.exist? document
-tester = Docspec::Tester.new document
-tester.execute
+document = Docspec::Document.from_file document
+document.test
 
 say ''
 
-if tester.success?
-  say "!txtgrn!#{tester.total} tests, #{tester.errors} failed\n"
+if document.success?
+  say "!txtgrn!#{document.total_examples} tests, #{document.failed_examples} failed\n"
   exit 0
 else
-  say "!txtred!#{tester.total} tests, #{tester.errors} failed\n"
+  say "!txtred!#{document.total_examples} tests, #{document.failed_examples} failed\n"
   exit 1
 end
