+++ lib/docspec.rb
@@ -2,5 +2,5 @@ require 'byebug' if ENV['BYEBUG']
 
 require 'docspec/output_capturer'
 require 'docspec/example'
-require 'docspec/tester'
+require 'docspec/testable'
 require 'docspec/document'
