+++ sample.md
@@ -1,0 +1,7 @@
+This file is used by one of the examples in the README.
+
+```ruby
+# Sample Test
+puts "sample test"
+#=> sample test
+```
